Move get_device to trainer.
