+++ src/continuity/data/__init__.py
@@ -7,14 +7,11 @@ Every data set is a list of `(x, u, y, v)` tuples.
 
 from .dataset import OperatorDataset
 from .shape import DatasetShapes
-from .utility import get_device, split, dataset_loss
+from .utility import split, dataset_loss
 
 __all__ = [
     "OperatorDataset",
     "DatasetShapes",
-    "get_device",
     "split",
     "dataset_loss",
 ]
-
-device = get_device()
